backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
bidict==0.23.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
dnspython==2.7.0
email-validator==2.2.0
fastapi==0.115.3
fastapi-cli==0.0.5
fastapi-login==1.10.2
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
itsdangerous==2.2.0
jinja2==3.1.4
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
mypy-extensions==1.0.0
orjson==3.10.10
packaging==24.1
pathspec==0.12.1
platformdirs==4.3.6
psycopg==3.2.3
psycopg-binary==3.2.3
pycparser==2.22
pydantic==2.9.2
pydantic-core==2.23.4
pydantic-extra-types==2.9.0
pydantic-settings==2.6.0
pygments==2.18.0
pyjwt==2.9.0
python-dotenv==1.0.1
python-engineio==4.10.1
python-multipart==0.0.12
pyyaml==6.0.2
requests==2.32.3
resend==2.4.0
rich==13.9.3
shellingham==1.5.4
simple-websocket==1.1.0
sniffio==1.3.1
sqlparse==0.5.1
starlette==0.41.0
typer==0.12.5
typing-extensions==4.12.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
watchfiles==0.24.0
websockets==14.1
wsproto==1.2.0
